requirements.txt
Appium-Python-Client==2.6.1;python_version >= '3.0'
browserstack-local==1.2.2
pytest==6.2.2
pytest-html==3.1.1
allure-pytest==2.13.0